revert: volver el firmware al estado previo a la fase 4
La telemetria dejo de bajar de 50 a 20 Hz al activarse el LiDAR y acabo
cayendo a 2 Hz. La instrumentacion de la fase 4 --comandos de control,
OTA, historial en RTC RAM, traza circular, watchdog de RX-- degrado el
bucle principal hasta el punto de no poder drenar el UART del LiDAR:

  loop_max_ms  34 ms (antes de la fase 4)
               680 ms (con la fase 4)

Limitar los sondeos UDP a 50 ms lo bajo a 10 ms pero no restauro el
comportamiento del LiDAR, asi que el problema no era solo la frecuencia
de sondeo: hay algo mas en esa instrumentacion que compite con la
lectura serial.

Se revierten src/, include/, lib/ y platformio.ini a b862970. Se
conservan la documentacion y los scripts, que no afectan al firmware. El
trabajo queda intacto en la rama fase4-instrumentacion para reintroducir
las piezas de a una, midiendo el LiDAR en cada paso.

Verificado en el robot tras cargar por OTA (la ultima carga posible por
WiFi, ya que la version revertida no lo incluye): loop_max_ms 34 ms y el
reporte sin los campos de OTA ni de sesion.

+++ lib/motor_ctl_kalman/src/motor_ctl_kalman.cpp
@@ -198,8 +198,8 @@ void MotorController::update() {
 
   tickSampleTimePrev = tickTime;
 
-  int32_t encNow = getEncoderValue();
-  int32_t encDelta = encNow - encPrev;
+  long int encNow = getEncoderValue();
+  long int encDelta = encNow - encPrev;
   encPrev = encNow;
   float ticksPerMicroSec = ((float) encDelta) / ((float) tickTimeDelta);
   measuredRPM = ticksPerMicroSec * ticksPerMicroSecToRPM;
